Decouple ast into expr and statement.
